feat(api)!: switch URL path version segment from full SemVer to major-only
The URL path segment changes from `/api/3.1.0/...` to `/api/v4/...`.
Route handlers move from `src/routes/{public,private,delete}/` to
`src/routes/v4/{...}`. The directory name carries the version literal;
references to `v4` are hardcoded at each callsite rather than threaded
through a constant, since the directory rename and the callsite update
move together when bumping majors.

`apiVersion` field removed from `version.json`; `getApiVersion()` and
`API_VERSION` removed from `src/config.ts`. The package `version` and
`docVersion` fields stay.

BREAKING CHANGE: Existing clients that hit `/api/3.1.0/<endpoint>` will
get 404. Update your base URL to `/api/v4/<endpoint>`.

Closes #114
