+++ src/routes/index.ts
@@ -1,13 +1,6 @@
-import express, { Router } from 'express';
-import path from 'path';
-import { publicRouter } from './public';
-import { privateRouter } from './private';
-import { deleteRouter } from './delete';
-import { LOCAL_DIRECTORY, __dirname } from '../config';
+import { Router } from 'express';
+import { v4Router } from './v4';
 
 export const router = Router();
 
-router.use('/public', publicRouter);
-router.use('/private', privateRouter);
-router.use('/', deleteRouter);
-router.use(express.static(path.join(__dirname, LOCAL_DIRECTORY)));
+router.use('/v4', v4Router);
