+++ src/routes/v4/private/controller.test.ts
@@ -1,4 +1,3 @@
-import fs from 'fs';
 import os from 'os';
 import path from 'path';
 import { getMockReq, getMockRes } from '@jest-mock/express';
@@ -12,7 +11,7 @@ const mockLogger: any = {
 };
 mockLogger.child = jest.fn(() => mockLogger);
 
-jest.mock('../../services/logging', () => ({
+jest.mock('../../../services/logging', () => ({
     getLogger: () => mockLogger,
     apiLogger: mockLogger,
     authLogger: mockLogger,
@@ -22,77 +21,86 @@ jest.mock('../../services/logging', () => ({
     serverLogger: mockLogger,
 }));
 
-import { storePublic } from './controller';
-import { PublicService } from './service';
-import { BadRequestError } from '../../errors';
+import { BadRequestError } from '../../../errors';
+import { PrivateService } from './service';
 
 const { res: mockRes, next: mockNext, clearMockRes } = getMockRes();
 
-jest.mock('../../config', () => ({
+jest.mock('../../../config', () => ({
     AVAILABLE_BUCKETS: ['bucketName'],
     STORAGE_TYPE: 'gcp',
 }));
 
-jest.mock('../../services/storage/gcp', () => ({
+jest.mock('../../../services/cryptography', () => ({
+    CryptographyService: jest.fn().mockImplementation(() => ({
+        computeDigestMultibase: jest.fn().mockResolvedValue('mocked-digest'),
+        generateEncryptionKey: jest.fn().mockResolvedValue('test-encryption-key'),
+        encryptString: jest
+            .fn()
+            .mockReturnValue({ cipherText: 'encrypted', iv: 'test-iv', tag: 'test-tag', type: 'aes-256-gcm' }),
+    })),
+}));
+
+jest.mock('../../../services/storage/gcp', () => ({
     GCPStorageService: jest.fn().mockImplementation(() => ({
         uploadFile: jest.fn().mockResolvedValue({ uri: 'mock-uri' }),
         objectExists: jest.fn().mockResolvedValue(false),
     })),
 }));
 
-jest.mock('../../services/cryptography', () => ({
-    CryptographyService: jest.fn().mockImplementation(() => ({
-        computeDigestMultibase: jest.fn().mockResolvedValue('mocked-digest'),
-    })),
-}));
-
 jest.mock('fs', () => ({
     ...jest.requireActual('fs'),
     promises: {
-        readFile: jest.fn().mockResolvedValue(Buffer.from('fake-binary-content')),
+        readFile: jest.fn().mockResolvedValue(Buffer.from('binary-content')),
         unlink: jest.fn().mockResolvedValue(undefined),
     },
 }));
 
-describe('PublicController', () => {
+import fs from 'fs';
+import { storePrivate } from './controller';
+
+describe('Private Controller', () => {
     beforeEach(() => {
         clearMockRes();
         jest.clearAllMocks();
     });
 
-    describe('storePublic', () => {
-        it('should successfully store a JSON document and return 201 with uri and digestMultibase', async () => {
-            const storeDocumentSpy = jest.spyOn(PublicService.prototype, 'storeDocument').mockResolvedValue({
+    describe('storePrivate', () => {
+        it('should successfully store a JSON document and return 201', async () => {
+            const spy = jest.spyOn(PrivateService.prototype, 'encryptAndStoreDocument').mockResolvedValue({
                 uri: 'mock-uri',
                 digestMultibase: 'mocked-digest',
+                decryptionKey: 'test-encryption-key',
             });
 
             const mockReq = getMockReq({
                 body: {
                     bucket: 'bucketName',
-                    id: '123e4567-e89b-12d3-a456-426614174000',
-                    data: { name: 'test' },
+                    id: '550e8400-e29b-41d4-a716-446655440000',
+                    data: { key: 'value' },
                 },
             });
 
-            await storePublic(mockReq, mockRes, mockNext);
+            await storePrivate(mockReq, mockRes, mockNext);
 
             expect(mockRes.status).toHaveBeenCalledWith(201);
             expect(mockRes.json).toHaveBeenCalledWith({
                 uri: 'mock-uri',
                 digestMultibase: 'mocked-digest',
+                decryptionKey: 'test-encryption-key',
             });
-            expect(storeDocumentSpy).toHaveBeenCalledWith(
+            expect(spy).toHaveBeenCalledWith(
                 expect.anything(), // storageService
-                expect.anything(), // cryptoService
-                { bucket: 'bucketName', id: '123e4567-e89b-12d3-a456-426614174000', data: { name: 'test' } },
+                expect.anything(), // cryptographyService
+                { bucket: 'bucketName', id: '550e8400-e29b-41d4-a716-446655440000', data: { key: 'value' } },
             );
         });
 
-        it('should successfully store a binary file and return 201 with uri and digestMultibase', async () => {
-            const storeFileSpy = jest.spyOn(PublicService.prototype, 'storeFile').mockResolvedValue({
-                uri: 'mock-file-uri',
-                digestMultibase: 'mocked-file-digest',
+        it('should successfully store a binary file and return 201', async () => {
+            const spy = jest.spyOn(PrivateService.prototype, 'encryptAndStoreFile').mockResolvedValue({
+                uri: 'mock-uri',
+                digestMultibase: 'mocked-digest',
+                decryptionKey: 'test-encryption-key',
             });
 
             const tempFilePath = path.join(os.tmpdir(), 'upload-12345');
@@ -100,27 +108,28 @@ describe('PublicController', () => {
                 file: {
                     path: tempFilePath,
                     mimetype: 'image/png',
-                } as Express.Multer.File,
+                } as any,
                 body: {
                     bucket: 'bucketName',
-                    id: '123e4567-e89b-12d3-a456-426614174000',
+                    id: '550e8400-e29b-41d4-a716-446655440000',
                 },
             });
 
-            await storePublic(mockReq, mockRes, mockNext);
+            await storePrivate(mockReq, mockRes, mockNext);
 
             expect(fs.promises.readFile).toHaveBeenCalledWith(path.resolve(tempFilePath));
             expect(mockRes.status).toHaveBeenCalledWith(201);
             expect(mockRes.json).toHaveBeenCalledWith({
-                uri: 'mock-file-uri',
-                digestMultibase: 'mocked-file-digest',
+                uri: 'mock-uri',
+                digestMultibase: 'mocked-digest',
+                decryptionKey: 'test-encryption-key',
             });
-            expect(storeFileSpy).toHaveBeenCalledWith(
-                expect.anything(), // storageService
-                expect.anything(), // cryptoService
+            expect(spy).toHaveBeenCalledWith(
+                expect.anything(),
+                expect.anything(),
                 expect.objectContaining({
                     bucket: 'bucketName',
-                    id: '123e4567-e89b-12d3-a456-426614174000',
+                    id: '550e8400-e29b-41d4-a716-446655440000',
                     file: expect.any(Buffer),
                     mimeType: 'image/png',
                 }),
@@ -128,18 +137,18 @@ describe('PublicController', () => {
         });
 
         it('should return 400 for an invalid bucket', async () => {
-            jest.spyOn(PublicService.prototype, 'storeDocument').mockRejectedValue(
+            jest.spyOn(PrivateService.prototype, 'encryptAndStoreDocument').mockRejectedValue(
                 new BadRequestError('Invalid bucket. Must be one of the following buckets: bucketName'),
             );
 
             const mockReq = getMockReq({
                 body: {
-                    bucket: 'invalid-bucket',
-                    data: { name: 'test' },
+                    bucket: 'invalidBucket',
+                    data: { key: 'value' },
                 },
             });
 
-            await storePublic(mockReq, mockRes, mockNext);
+            await storePrivate(mockReq, mockRes, mockNext);
 
             expect(mockRes.status).toHaveBeenCalledWith(400);
             expect(mockRes.json).toHaveBeenCalledWith({
@@ -147,23 +156,23 @@ describe('PublicController', () => {
             });
         });
 
-        it('should return 500 for unexpected errors', async () => {
-            jest.spyOn(PublicService.prototype, 'storeDocument').mockRejectedValue(
+        it('should return 500 when a non-ApiError is thrown', async () => {
+            jest.spyOn(PrivateService.prototype, 'encryptAndStoreDocument').mockRejectedValue(
                 new Error('Something went terribly wrong'),
             );
 
             const mockReq = getMockReq({
                 body: {
                     bucket: 'bucketName',
-                    data: { name: 'test' },
+                    data: { key: 'value' },
                 },
             });
 
-            await storePublic(mockReq, mockRes, mockNext);
+            await storePrivate(mockReq, mockRes, mockNext);
 
             expect(mockRes.status).toHaveBeenCalledWith(500);
             expect(mockRes.json).toHaveBeenCalledWith({
-                message: 'An unexpected error occurred while storing the resource.',
+                message: 'An unexpected error occurred while storing private data.',
             });
         });
 
@@ -171,7 +180,7 @@ describe('PublicController', () => {
             const mockReq = getMockReq({
                 body: {
                     bucket: 'bucketName',
-                    id: '123e4567-e89b-12d3-a456-426614174000',
+                    id: '550e8400-e29b-41d4-a716-446655440000',
                 },
             });
             mockReq.is = jest.fn((type: string) =>
@@ -179,7 +188,7 @@ describe('PublicController', () => {
             ) as any;
             mockReq.file = undefined;
 
-            await storePublic(mockReq, mockRes, mockNext);
+            await storePrivate(mockReq, mockRes, mockNext);
 
             expect(mockRes.status).toHaveBeenCalledWith(400);
             expect(mockRes.json).toHaveBeenCalledWith({
@@ -188,41 +197,45 @@ describe('PublicController', () => {
         });
 
         it('should clean up the temporary file after a binary upload', async () => {
-            jest.spyOn(PublicService.prototype, 'storeFile').mockResolvedValue({
-                uri: 'mock-file-uri',
-                digestMultibase: 'mocked-file-digest',
-            });
-
-            const tempFilePath = path.join(os.tmpdir(), 'upload-cleanup-test');
-            const mockReq = getMockReq({
-                file: {
-                    path: tempFilePath,
-                    mimetype: 'image/png',
-                } as Express.Multer.File,
-                body: {
-                    bucket: 'bucketName',
-                },
-            });
-
-            await storePublic(mockReq, mockRes, mockNext);
-
-            expect(fs.promises.unlink).toHaveBeenCalledWith(path.resolve(tempFilePath));
-        });
-        it('should not attempt to clean up when there is no temporary file', async () => {
-            jest.spyOn(PublicService.prototype, 'storeDocument').mockResolvedValue({
+            jest.spyOn(PrivateService.prototype, 'encryptAndStoreFile').mockResolvedValue({
                 uri: 'mock-uri',
                 digestMultibase: 'mocked-digest',
+                decryptionKey: 'test-encryption-key',
+            });
+
+            const tempPath = path.join(os.tmpdir(), 'upload-cleanup-test');
+            const mockReq = getMockReq({
+                file: {
+                    path: tempPath,
+                    mimetype: 'image/png',
+                } as any,
+                body: {
+                    bucket: 'bucketName',
+                    id: '550e8400-e29b-41d4-a716-446655440000',
+                },
+            });
+
+            await storePrivate(mockReq, mockRes, mockNext);
+
+            expect(fs.promises.unlink).toHaveBeenCalledWith(path.resolve(tempPath));
+        });
+
+        it('should not attempt to clean up when there is no temporary file', async () => {
+            jest.spyOn(PrivateService.prototype, 'encryptAndStoreDocument').mockResolvedValue({
+                uri: 'mock-uri',
+                digestMultibase: 'mocked-digest',
+                decryptionKey: 'test-encryption-key',
             });
 
             const mockReq = getMockReq({
                 body: {
                     bucket: 'bucketName',
-                    id: '123e4567-e89b-12d3-a456-426614174000',
-                    data: { name: 'test' },
+                    id: '550e8400-e29b-41d4-a716-446655440000',
+                    data: { key: 'value' },
                 },
             });
 
-            await storePublic(mockReq, mockRes, mockNext);
+            await storePrivate(mockReq, mockRes, mockNext);
 
             expect(fs.promises.unlink).not.toHaveBeenCalled();
         });
@@ -235,13 +248,13 @@ describe('PublicController', () => {
                 file: {
                     path: tempPath,
                     mimetype: 'image/png',
-                } as Express.Multer.File,
+                } as any,
                 body: {
                     bucket: 'bucketName',
                 },
             });
 
-            await storePublic(mockReq, mockRes, mockNext);
+            await storePrivate(mockReq, mockRes, mockNext);
 
             expect(fs.promises.unlink).toHaveBeenCalledWith(path.resolve(tempPath));
         });
@@ -251,9 +264,10 @@ describe('PublicController', () => {
                 Object.assign(new Error('EACCES: permission denied'), { code: 'EACCES' }),
             );
 
-            jest.spyOn(PublicService.prototype, 'storeFile').mockResolvedValue({
+            jest.spyOn(PrivateService.prototype, 'encryptAndStoreFile').mockResolvedValue({
                 uri: 'mock-uri',
                 digestMultibase: 'mocked-digest',
+                decryptionKey: 'test-encryption-key',
             });
 
             const tempPath = path.join(os.tmpdir(), 'upload-unlink-error');
@@ -262,7 +276,7 @@ describe('PublicController', () => {
                 body: { bucket: 'bucketName', id: '550e8400-e29b-41d4-a716-446655440000' },
             });
 
-            await storePublic(mockReq, mockRes, mockNext);
+            await storePrivate(mockReq, mockRes, mockNext);
 
             expect(fs.promises.unlink).toHaveBeenCalledWith(path.resolve(tempPath));
             expect(mockLogger.error).toHaveBeenCalledWith(
@@ -276,14 +290,14 @@ describe('PublicController', () => {
                 file: {
                     path: '/etc/passwd',
                     mimetype: 'application/octet-stream',
-                } as Express.Multer.File,
+                } as any,
                 body: {
                     bucket: 'bucketName',
-                    id: '123e4567-e89b-12d3-a456-426614174000',
+                    id: '550e8400-e29b-41d4-a716-446655440000',
                 },
             });
 
-            await storePublic(mockReq, mockRes, mockNext);
+            await storePrivate(mockReq, mockRes, mockNext);
 
             expect(mockRes.status).toHaveBeenCalledWith(400);
             expect(mockRes.json).toHaveBeenCalledWith({
