+++ src/routes/v4/private/index.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express';
 import { storePrivate } from './controller';
-import { authenticateRequest } from '../../middleware/authentication';
-import { conditionalUpload, handleUploadError } from '../../middleware/upload';
+import { authenticateRequest } from '../../../middleware/authentication';
+import { conditionalUpload, handleUploadError } from '../../../middleware/upload';
 
 export const privateRouter = Router();
