+++ src/services/storage/local.test.ts
@@ -1,18 +1,12 @@
 import { LocalStorageService } from './local';
 
-const { apiVersion: API_VERSION } = require('../../../version.json');
-
-jest.mock('../../config', () => {
-    const { apiVersion } = require('../../../version.json');
-    return {
-        API_VERSION: apiVersion,
-        DOMAIN: 'localhost',
-        LOCAL_DIRECTORY: 'uploads',
-        PORT: '3333',
-        EXTERNAL_PORT: '3333',
-        PROTOCOL: 'http',
-    };
-});
+jest.mock('../../config', () => ({
+    DOMAIN: 'localhost',
+    LOCAL_DIRECTORY: 'uploads',
+    PORT: '3333',
+    EXTERNAL_PORT: '3333',
+    PROTOCOL: 'http',
+}));
 
 jest.mock('fs', () => {
     const actual = jest.createMockFromModule<typeof import('fs')>('fs');
@@ -41,7 +35,7 @@ describe('LocalStorageService', () => {
 
         const result = await storageService.uploadFile(bucket, key, body, contentType);
 
-        expect(result.uri).toEqual(`http://localhost:3333/api/${API_VERSION}/test-bucket/test-file.json`);
+        expect(result.uri).toEqual(`http://localhost:3333/api/v4/test-bucket/test-file.json`);
     });
 
     it('should upload a Buffer body to the local file system', async () => {
@@ -52,7 +46,7 @@ describe('LocalStorageService', () => {
 
         const result = await storageService.uploadFile(bucket, key, body, contentType);
 
-        expect(result.uri).toEqual(`http://localhost:3333/api/${API_VERSION}/test-bucket/test-file.bin`);
+        expect(result.uri).toEqual(`http://localhost:3333/api/v4/test-bucket/test-file.bin`);
     });
 
     it('should use key as-passed in objectExists without appending .json', () => {
