+++ version.json
@@ -1,6 +1,5 @@
 {
     "version": "3.2.1",
-    "apiVersion": "3.1.0",
     "docVersion": "3.2.0",
     "dependencies": {}
 }
